+++ README.md
@@ -39,18 +39,7 @@ This will create a Project with already configured guardrails and evaluations fo
     hatch run create-vector-database
     ```
 
-3a. Run the agent (CLI version):
-
-    ```bash
-    hatch run agent --validation-mode cleanlab_log_tools
-    ```
-
-    **Note:** Use `--validation-mode` to control validation: `none` (default, no validation), `cleanlab` (standard validation), or `cleanlab_log_tools` (validation with post-chat-turn tool logging)
-
-
-3b. Run the agent (UI version):
-
-Run the backend server:
+3. Run the backend server:
 
     ```bash
     hatch run backend-server
@@ -58,7 +47,7 @@ Run the backend server:
 
     **Note:** This will run the backend server with cleanlab validation enabled by default. Include the `--disable-cleanlab` flag to disable validation.
 
-Run the frontend:
+4. Run the frontend:
 
     ```bash
     cd frontend
@@ -73,7 +62,7 @@ Run the frontend:
 
 This demo AI is a multi-turn conversational customer support agent for an airline, powered by RAG retrieval with responses generated by OpenAI's GPT-4o model. For realism, this demo uses real data scraped from Frontier Airlines' website, but can be replaced by any topic of any customer support agent. The agent uses specialized tools to search airline policy documents, retrieve articles, browse directories, and find the cheapest available flight deals across routes in natural conversations. Built with pydantic-ai, this system is connected to a [Cleanlab Project](https://codex.cleanlab.ai/) with advanced validation features including detailed tool call logging for complete transparency into how the agent retrieves information and constructs responses.
 
-**Temporary (only for now):** Go through this walkthrough with one window running two terminals with airline agents: one connected to a [Cleanlab Project](https://codex.cleanlab.ai/) using `--validation-mode cleanlab_log_tools`, one not; and another window showing the connected [Cleanlab Project](https://codex.cleanlab.ai/) dashboard.
+**Temporary (only for now):** Go through this walkthrough with one window running the airline agent UI backed by Cleanlab validation and another window showing the connected [Cleanlab Project](https://codex.cleanlab.ai/) dashboard. To compare responses with and without Cleanlab, restart the backend using the `--disable-cleanlab` flag.
 
 ### 1. Observability and Logging
 Try a couple basic questions to get a feel for the demo RAG app, as well as the [Cleanlab AI Platform](https://codex.cleanlab.ai/projects). Here are some ideas for questions you can ask.
